[Refactor]: user reader, store 패키지 변경(port)
