--- zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/user/UserStore.java
@@ -1,5 +1,0 @@
-package org.haedal.zzansuni.domain.user;
-
-public interface UserStore {
-    User store(User user);
-}
