+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/user/port/UserStore.java
@@ -1,0 +1,7 @@
+package org.haedal.zzansuni.domain.user.port;
+
+import org.haedal.zzansuni.domain.user.User;
+
+public interface UserStore {
+    User store(User user);
+}
