+++ zzansuni-api-server/app/src/test/java/org/haedal/zzansuni/domain/user/UserServiceTest.java
@@ -1,5 +1,6 @@
 package org.haedal.zzansuni.domain.user;
 
+import org.haedal.zzansuni.domain.user.port.UserReader;
 import org.haedal.zzansuni.global.security.Role;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,7 +8,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
